+++ src/constants/appVersion.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = "13.137.12";
+export const APP_VERSION = "13.137.13";